if文を一行で記述

+++ app/controllers/items_controller.rb
@@ -48,9 +48,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if (current_user.id == item.user_id)
-      item.destroy
-    end
+    item.destroy if current_user.id == item.user_id
   end
 
   def show
